+++ lib/papierkram_api/v1/endpoints/expense/vouchers.rb
@@ -38,6 +38,147 @@ module PapierkramApi
 
             http_get("#{@url_api_path}/expense/vouchers", query)
           end
+
+          def create( # rubocop:disable Metrics/ParameterLists,Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/PerceivedComplexity,Metrics/MethodLength
+            name:,
+            provenance:,
+            line_items:,
+            due_date: nil,
+            document_date: nil,
+            description: nil,
+            entertainment_reason: nil,
+            entertainment_persons: nil,
+            flagged: nil,
+            creditor_id: nil
+          )
+            body = {}
+            body[:name] = name
+            body[:line_items] = line_items
+            body[:provenance] = provenance
+
+            body[:due_date] = due_date if due_date
+            body[:document_date] = document_date if document_date
+            body[:description] = description if description
+            body[:entertainment_reason] = entertainment_reason if entertainment_reason
+            body[:entertainment_persons] = entertainment_persons if entertainment_persons
+            body[:flagged] = flagged if flagged
+            body[:creditor] = { id: creditor_id } if creditor_id
+
+            if entertainment_persons
+              raise ArgumentError, 'entertainment_persons must be an array' unless entertainment_persons.is_a?(Array)
+
+              unless entertainment_persons.all?(String)
+                raise ArgumentError,
+                      'entertainment_persons must be an array of strings'
+              end
+            end
+
+            if line_items
+              raise ArgumentError, 'line_items must be an array' unless line_items.is_a?(Array)
+              raise ArgumentError, 'line_items must be an array of hashes' unless line_items.all?(Hash)
+            end
+
+            http_post("#{@url_api_path}/expense/vouchers", body)
+          end
+
+          def update_by( # rubocop:disable Metrics/ParameterLists,Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/PerceivedComplexity
+            id:,
+            name: nil,
+            due_date: nil,
+            document_date: nil,
+            description: nil,
+            entertainment_reason: nil,
+            entertainment_persons: nil,
+            flagged: nil,
+            provenance: nil,
+            creditor_id: nil,
+            line_items: nil
+          )
+            body = {}
+            body[:name] = name if name
+            body[:due_date] = due_date if due_date
+            body[:document_date] = document_date if document_date
+            body[:description] = description if description
+            body[:entertainment_reason] = entertainment_reason if entertainment_reason
+            body[:entertainment_persons] = entertainment_persons if entertainment_persons
+            body[:flagged] = flagged if flagged
+            body[:provenance] = provenance if provenance
+            body[:creditor] = { id: creditor_id } if creditor_id
+            body[:line_items] = line_items if line_items
+
+            if line_items
+              raise ArgumentError, 'line_items must be an array' unless line_items.is_a?(Array)
+              raise ArgumentError, 'line_items must be an array of hashes' unless line_items.all?(Hash)
+            end
+
+            http_put("#{@url_api_path}/expense/vouchers/#{id}", body)
+          end
+
+          def add_document_by(id:, path_to_file:, file_type:)
+            doc = Faraday::Multipart::FilePart.new(path_to_file, file_type)
+
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/documents", { file: doc })
+          end
+
+          def delete_document_by(id:, document_id:)
+            http_delete("#{@url_api_path}/expense/vouchers/#{id}/documents/#{document_id}")
+          end
+
+          def delete_by(id:)
+            http_delete("#{@url_api_path}/expense/vouchers/#{id}")
+          end
+
+          def archive_by(id:)
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/archive")
+          end
+
+          def unarchive_by(id:)
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/unarchive")
+          end
+
+          def cancel_by(id:)
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/cancel")
+          end
+
+          def cancel_with_reverse_entry_by(id:)
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/cancel_with_reverse_entry")
+          end
+
+          def mark_as_paid_by( # rubocop:disable Metrics/CyclomaticComplexity,Metrics/MethodLength,Metrics/PerceivedComplexity
+            id:,
+            value: nil,
+            payment_date: nil,
+            banking_transaction_id: nil,
+            difference_reason: nil
+          )
+
+            unless value || banking_transaction_id
+              raise ArgumentError,
+                    'either value or banking_transaction_id must be given'
+            end
+
+            allowed_difference_reasons = %w[sonstige mahnung teilzahlung skonto sonstige schmaelerung]
+            body = {}
+            body[:payment_date] = payment_date if payment_date
+            if value
+              raise ArgumentError, 'value must be a float' unless value.is_a?(Numeric)
+
+              body[:value] = value
+            end
+            body[:banking_transaction] = { id: banking_transaction_id } if banking_transaction_id
+
+            if difference_reason
+              raise ArgumentError, 'difference_reason must be a string' unless difference_reason.is_a?(String)
+
+              unless allowed_difference_reasons.include?(difference_reason)
+                raise ArgumentError, "difference_reason must be one of: #{allowed_difference_reasons.join(', ')}"
+              end
+
+              body[:difference_reason] = difference_reason
+            end
+
+            http_post("#{@url_api_path}/expense/vouchers/#{id}/pay", body)
+          end
         end
       end
     end